worker,types: refactor: single failure classification module and a transient wire default
